Move collaborator mode to editor store, and remove the temporary block on core-data actions
