+++ packages/editor/src/store/actions.js
@@ -667,6 +667,20 @@ export function setDeviceType( deviceType ) {
 }
 
 /**
+ * Returns an action object used in signalling that the collaborator mode has been set.
+ *
+ * @param {'view' | 'edit'} collaboratorMode The collaborator mode.
+ *
+ * @return {Object} Action object.
+ */
+export function setCollaboratorMode( collaboratorMode ) {
+	return {
+		type: 'SET_COLLABORATOR_MODE',
+		collaboratorMode,
+	};
+}
+
+/**
  * Returns an action object used to enable or disable a panel in the editor.
  *
  * @param {string} panelName A string that identifies the panel to enable or disable.
